# specs/001-todo-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TODO API
  version: 1.0.0
  description: API HTTP para gerenciar tarefas TODO.
paths:
  /tasks:
    post:
      summary: Criar tarefa
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskInput'
      responses:
        '201':
          description: Tarefa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
    get:
      summary: Listar tarefas
      operationId: listTasks
      parameters:
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum:
              - priority
        - name: sortOrder
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum:
              - asc
              - desc
      responses:
        '200':
          description: Lista de tarefas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
  /tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      summary: Buscar tarefa por identificador
      operationId: getTask
      responses:
        '200':
          description: Tarefa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/TaskNotFound'
    patch:
      summary: Atualizar titulo ou descricao
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskInput'
      responses:
        '200':
          description: Tarefa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/TaskNotFound'
    delete:
      summary: Remover tarefa
      operationId: deleteTask
      responses:
        '204':
          description: Tarefa removida
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/TaskNotFound'
  /tasks/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    patch:
      summary: Concluir tarefa
      operationId: completeTask
      responses:
        '200':
          description: Tarefa concluida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/TaskNotFound'
components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Entrada invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TaskNotFound:
      description: Tarefa nao encontrada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          type: string
          enum:
            - pending
            - completed
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTaskInput:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
    UpdateTaskInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
        description:
          type:
            - string
            - 'null'
        priority:
          $ref: '#/components/schemas/TaskPriority'
    TaskPriority:
      type: integer
      format: int32
      description: Prioridade da tarefa. 1 = high, 2 = medium, 3 = low.
      enum:
        - 1
        - 2
        - 3
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          examples:
            - VALIDATION_ERROR
            - TASK_NOT_FOUND
        message:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: true